+++ pkgs/artisanal_widgets/test/components/frame_left_border_test.dart
@@ -1,6 +1,6 @@
 import 'package:artisanal/style.dart' as style;
-import 'package:artisanal_widgets/artisanal_widgets.dart' as w;
-import 'package:artisanal_widgets/testing.dart';
+import 'package:artisanal/testing.dart';
+import 'package:artisanal/widgets.dart' as w;
 import 'package:test/test.dart';
 
 void main() {